~~ please do a gradle build before you push ~~
the code currently doesn't pass a test, so whoever broke
it needs to fix it soon! the test might be wrong, fyi.

this commit is still broken, :(

anyway, the GameEngine is not an object to be created,
more of a library with useful functions.

also, the arguments passed to Game are of the form
<host> <port> ...

so there are actually twice as many args as there will be players.