fixed bug in Game when first player is booted
this bug was caused by cycling the linked list of players by the total number
of players (not the number of remaining players). if player 0 was booted in
before the MOVE message, the client would still cycle the list 2 or 4 times,
causing the first player to not be player 0 and all kinds of chaos would
ensue with the GameEngine, the Messenger, the Frame, and so on.